BUG: revert division change in downsample code
Previous update mistakenly converted a division from float
to integer where float was actually intended and required.

#717e4e71

Other statements (from future) in that commit could be removed
but they are harmless.

+++ Modules/Scripted/TractographyDownsample/TractographyDownsample.py
@@ -325,7 +325,7 @@ class TractographyDownsampleLogic(ScriptedLoadableModuleLogic):
         point1 = inpoints.GetPoint(ptids.GetId(ptidx + 1))
         step_size += np.sqrt(np.sum(np.power(np.subtract(point0, point1), 2)))
         count += 1
-    step_size = int(step_size / count)
+    step_size = step_size / count
     return step_size
 
   def downsampleFibers(self, inpd, outpd, outstep, outpercent, outminpts, outminlen, outmaxlen):
